requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-groq
chromadb
sentence-transformers
transformers
torch
regex